src/main/java/by/afinny/insuranceservice/integration/NewCarPolicySours.java
package by.afinny.insuranceservice.integration;

import by.afinny.insuranceservice.dto.kafka.ProducerNewPolicyEvent;
import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.beans.factory.annotation.Value;
import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
import org.springframework.context.event.EventListener;
import org.springframework.kafka.core.KafkaTemplate;
import org.springframework.kafka.support.KafkaHeaders;
import org.springframework.messaging.MessageHeaders;
import org.springframework.messaging.support.MessageBuilder;
import org.springframework.stereotype.Component;
import org.springframework.util.MimeTypeUtils;

@Component
@RequiredArgsConstructor
@Slf4j
@ConditionalOnProperty("kafka.topics.new-policy-producer.enabled")
public class NewCarPolicySours {

    private final KafkaTemplate<String, ?> kafkaTemplate;

    @Value("${kafka.topics.new-policy-producer.path}")
    private String kafkaTopic;

    @EventListener
    public void sendMessageAboutPolicyInformation(ProducerNewPolicyEvent event) {
        log.info("Event " + event + " has been received, sending message...");
        kafkaTemplate.send(
                MessageBuilder
                        .withPayload(event)
                        .setHeader(MessageHeaders.CONTENT_TYPE, MimeTypeUtils.APPLICATION_JSON)
                        .setHeader(KafkaHeaders.TOPIC, kafkaTopic)
                        .build()
        );
    }
}
